refactor: use barrel pattern for hooks in Article component
