+++ src/components/article/hooks/index.ts
@@ -1,0 +1,5 @@
+import { useHandleDeleteArticle } from './useHandleDeleteArticle';
+import { useHandleDeleteComment } from './useHandleDeleteComment';
+import { uesHandlePostComment } from './useHandlePostComment';
+
+export { useHandleDeleteArticle, useHandleDeleteComment, uesHandlePostComment };
